add casts

+++ include/stencil-composition/structured_grids/backend_cuda/execute_kernel_functor_cuda.hpp
@@ -142,9 +142,9 @@ namespace gridtools {
             int jblock = max_extent_t::jminus::value - 1;
             if (threadIdx.y < jboundary_limit) {
                 i = blockIdx.x * block_size_t::i_size_t::value + threadIdx.x;
-                j = blockIdx.y * block_size_t::j_size_t::value + threadIdx.y + max_extent_t::jminus::value;
+                j = (int)blockIdx.y * block_size_t::j_size_t::value + (int)threadIdx.y + max_extent_t::jminus::value;
                 iblock = threadIdx.x;
-                jblock = threadIdx.y + max_extent_t::jminus::value;
+                jblock = (int)threadIdx.y + max_extent_t::jminus::value;
             } else if (threadIdx.y < iminus_limit) {
                 const int padded_boundary_ = padded_boundary< -max_extent_t::iminus::value >::value;
                 // we dedicate one warp to execute regions (a,h,e), so here we make sure we have enough threads
@@ -153,10 +153,10 @@ namespace gridtools {
                        enumtype::vector_width);
 
                 i = blockIdx.x * block_size_t::i_size_t::value - padded_boundary_ + threadIdx.x % padded_boundary_;
-                j = blockIdx.y * block_size_t::j_size_t::value + threadIdx.x / padded_boundary_ +
+                j = (int)blockIdx.y * block_size_t::j_size_t::value + (int)threadIdx.x / padded_boundary_ +
                     max_extent_t::jminus::value;
                 iblock = -padded_boundary_ + (int)threadIdx.x % padded_boundary_;
-                jblock = threadIdx.x / padded_boundary_ + max_extent_t::jminus::value;
+                jblock = (int)threadIdx.x / padded_boundary_ + max_extent_t::jminus::value;
             } else if (threadIdx.y < iplus_limit) {
                 const int padded_boundary_ = padded_boundary< max_extent_t::iplus::value >::value;
                 // we dedicate one warp to execute regions (c,i,g), so here we make sure we have enough threads
@@ -166,10 +166,10 @@ namespace gridtools {
 
                 i = blockIdx.x * block_size_t::i_size_t::value + threadIdx.x % padded_boundary_ +
                     block_size_t::i_size_t::value;
-                j = blockIdx.y * block_size_t::j_size_t::value + threadIdx.x / padded_boundary_ +
+                j = (int)blockIdx.y * block_size_t::j_size_t::value + (int)threadIdx.x / padded_boundary_ +
                     max_extent_t::jminus::value;
                 iblock = threadIdx.x % padded_boundary_ + block_size_t::i_size_t::value;
-                jblock = threadIdx.x / padded_boundary_ + max_extent_t::jminus::value;
+                jblock = (int)threadIdx.x / padded_boundary_ + max_extent_t::jminus::value;
             }
             it_domain.set_index(0);
 
@@ -319,7 +319,7 @@ namespace gridtools {
 #endif
 
                 _impl_strcuda::do_it_on_gpu< run_functor_arguments_cuda_t,
-                    local_domain_t ><<< blocks, threads >>> //<<<nbx*nby, ntx*nty>>>
+                    local_domain_t ><<< blocks, threads >>>  //<<<nbx*nby, ntx*nty>>>
                     (local_domain_gp, grid_gp, m_grid.i_low_bound(), m_grid.j_low_bound(), (nx), (ny));
 
                 // TODOCOSUNA we do not need this. It will block the host, and we want to continue doing other stuff
